move `do_work` back into main function
