+++ src/git_client.rs
@@ -1,14 +1,11 @@
-use std::env::{current_dir, set_current_dir};
-use std::io::{self, BufRead, BufReader, Write};
-use std::process::{self, Command, Stdio};
+use std::io::{BufRead, BufReader};
+use std::process::{Command, Stdio};
 
-use crate::runners;
-
-const GIT: &str = "git";
-const WORKTREE_DIR: &str = "gitavs-worktree";
+pub const GIT: &str = "git";
+pub const WORKTREE_DIR: &str = "gitavs-worktree";
 
 #[derive(Default)]
-struct Commit {
+pub struct Commit {
     pub sha: String,
     pub message: String,
 }
@@ -31,72 +28,7 @@ impl Subcommand {
     }
 }
 
-pub fn do_work(from_sha: String) {
-    let mut cached_files: Vec<String> = vec![];
-    let repo_dir = current_dir().unwrap();
-
-    print!("\x1b[94m[GITAVS]\x1b[0m ⚙️ Setting up environment...");
-
-    create_worktree();
-
-    if set_current_dir(&format!("../{WORKTREE_DIR}").to_string()).is_err() {
-        eprintln!("Error, couldn't change to worktree directory");
-        delete_worktree();
-        process::exit(1);
-    }
-
-    runners::ruby::tests::rspec::setup_environment(repo_dir);
-
-    print!(" Done ✔️\n");
-    println!("\x1b[94m[GITAVS]\x1b[0m");
-
-    let commits: Vec<Commit> = get_commits(from_sha);
-
-    let mut i = 1;
-    for commit_pair in commits.windows(2) {
-        print!(
-            "\x1b[94m[GITAVS]\x1b[0m \x1b[33m{}\x1b[0m {}",
-            &commit_pair[1].sha, &commit_pair[1].message
-        );
-        print!(" ({i} of {})", commits.windows(2).len());
-
-        io::stdout().flush().expect("Failed to flush stdout");
-
-        let changed_files = get_changed_files(&commit_pair[0].sha, &commit_pair[1].sha);
-
-        if changed_files.is_empty() {
-            print!(" \x1b[2mNo test files\x1b[0m\n");
-
-            i += 1;
-
-            continue;
-        }
-
-        for file in &changed_files {
-            if !cached_files.contains(&file) {
-                cached_files.push(file.to_string());
-            }
-        }
-
-        checkout(&commit_pair[1].sha);
-
-        println!(
-            "\n\x1b[94m[GITAVS]\x1b[0m Changed files: {}",
-            changed_files.join(" ")
-        );
-        println!("\x1b[94m[GITAVS]\x1b[0m Running tests...");
-
-        runners::ruby::tests::rspec::run(&cached_files);
-
-        checkout(&"-".to_string());
-
-        i += 1;
-    }
-
-    delete_worktree();
-}
-
-fn get_commits(from_sha: String) -> Vec<Commit> {
+pub fn get_commits(from_sha: String) -> Vec<Commit> {
     let child = match Command::new(GIT)
         .args([
             Subcommand::Log.to_string(),
@@ -128,7 +60,7 @@ fn get_commits(from_sha: String) -> Vec<Commit> {
     }
 }
 
-fn get_changed_files(sha_1: &String, sha_2: &String) -> Vec<String> {
+pub fn get_changed_files(sha_1: &String, sha_2: &String) -> Vec<String> {
     let child = match Command::new(GIT)
         .args([Subcommand::Diff.to_string(), "--name-only", &sha_1, &sha_2])
         .stdout(Stdio::piped())
@@ -154,7 +86,7 @@ fn get_changed_files(sha_1: &String, sha_2: &String) -> Vec<String> {
     changed_files
 }
 
-fn checkout(value: &String) {
+pub fn checkout(value: &String) {
     match Command::new(GIT)
         .args([Subcommand::Checkout.to_string(), &format!("{}", value)])
         .output()
@@ -167,7 +99,7 @@ fn checkout(value: &String) {
     }
 }
 
-fn create_worktree() {
+pub fn create_worktree() {
     match Command::new(GIT)
         .args([
             Subcommand::Worktree.to_string(),
@@ -185,7 +117,7 @@ fn create_worktree() {
     }
 }
 
-fn delete_worktree() {
+pub fn delete_worktree() {
     match Command::new(GIT)
         .args([
             Subcommand::Worktree.to_string(),
